feat: add 10-day forecast component
The changes in this commit add a new component called ForecastDays. This component displays a 10-day forecast of the weather. It includes the day of the week, weather icon, and temperature range for each day. This component is used in multiple places within the application to provide users with an extended forecast.